getting data from Nasa api using different routers
